Added animation script file and css file and added some scroll animation class names to index.html.
Methodology is to add animation class to element in html, this class will set the initial changed state of the element.  Then the intersection observer will detect when the element is on screen and apply the active class to it, which will trigger the element to animate to its normal state.